refactor(E2E): Rename endpoints to __test_support__ pattern + update docs
🔐 E2E Endpoints Renamed (Platinum Compliance)

## Endpoint Changes
- OLD: /__e2e__/setup → NEW: /__test_support__/e2e/setup
- OLD: /__e2e__/cleanup → NEW: /__test_support__/e2e/cleanup

## Why This Change
- Term '__test_support__' is recognized (Rails, RSpec conventions)
- Clearly non-business and non-public
- Easy to blacklist in firewalls/proxies
- Security auditor friendly

## Security (Non-Negotiable)
- Routes ONLY exist when RAILS_ENV=test OR E2E_MODE=true
- Defense in depth: controller also checks e2e_mode_enabled?
- In production, routes don't exist (404)

## Documentation Updated
- README.md: Added E2E Testing section + Mission business rules
- docs/BRIEFING.md: Added E2E infrastructure details
- docs/index.md: Added E2E Testing Infrastructure section
- docs/technical/changes/2025-12-31-FC06_Missions_Implementation.md:
  - E2E section with endpoints and tests
  - Technical notes (CRA placeholder, post-WON notifications)

## Files Changed
- app/controllers/__test_support__/e2e/setup_controller.rb (moved + updated)
- config/routes.rb (new namespace pattern)
- bin/e2e/e2e_missions.sh (updated endpoints)
- .rubocop.yml (exclusion for test support controllers)
