+++ .rubocop.yml
@@ -47,6 +47,7 @@ Metrics/MethodLength:
     - "db/migrate/*"
     - "config/initializers/*"
     - "app/controllers/api/v1/missions_controller.rb"
+    - "app/controllers/__test_support__/**/*"
 
 # Metrics/AbcSize - augmenter la limite pour la complexité
 Metrics/AbcSize:
@@ -54,6 +55,7 @@ Metrics/AbcSize:
   Max: 28
   Exclude:
     - "app/controllers/api/v1/missions_controller.rb"
+    - "app/controllers/__test_support__/**/*"
 
 # Layout/LineLength - permettre des lignes plus longues
 Layout/LineLength:
