+++ config/routes.rb
@@ -28,12 +28,15 @@ Rails.application.routes.draw do
     end
   end
 
-  # E2E Setup Routes - ONLY mounted in test mode or when E2E_MODE=true
+  # E2E Test Support Routes - ONLY mounted in test mode or when E2E_MODE=true
   # 🔐 SECURITY: These routes are NOT accessible in production
+  # ⚠️ Any exposure in production is a CRITICAL security flaw
   if Rails.env.test? || ENV['E2E_MODE'] == 'true'
-    scope '__e2e__' do
-      post 'setup', to: 'e2e/setup#create'
-      delete 'cleanup', to: 'e2e/setup#destroy'
+    namespace :__test_support__, path: '__test_support__' do
+      namespace :e2e, path: 'e2e' do
+        post 'setup', to: 'setup#create'
+        delete 'cleanup', to: 'setup#destroy'
+      end
     end
   end
